Тесты, тесты и еще раз тесты, и немного про утечки Отчет по 3 части.

+++ libglnsvpos/include/libglnsvpos/glnsvpos.h
@@ -6,6 +6,6 @@
 #include <string>
 #include <sstream>
 
-int glnsvpos(bool RK_valid, double h);
+uint64_t glnsvpos(bool RK_valid, double h);
 
 #endif /* #ifndef GLNSVPOS_H */
